# internal/transport/httpapi/apigen/api.gen.go
//go:build go1.22

// Package apigen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package apigen

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// Defines values for RoomCommandCommandKind.
const (
	CreateRoom  RoomCommandCommandKind = "create_room"
	PauseRoom   RoomCommandCommandKind = "pause_room"
	PostMessage RoomCommandCommandKind = "post_message"
	ResumeRoom  RoomCommandCommandKind = "resume_room"
)

// Valid indicates whether the value is a known member of the RoomCommandCommandKind enum.
func (e RoomCommandCommandKind) Valid() bool {
	switch e {
	case CreateRoom:
		return true
	case PauseRoom:
		return true
	case PostMessage:
		return true
	case ResumeRoom:
		return true
	default:
		return false
	}
}

// Defines values for GetHealthz200JSONResponseBodyStatus.
const (
	Ok GetHealthz200JSONResponseBodyStatus = "ok"
)

// Valid indicates whether the value is a known member of the GetHealthz200JSONResponseBodyStatus enum.
func (e GetHealthz200JSONResponseBodyStatus) Valid() bool {
	switch e {
	case Ok:
		return true
	default:
		return false
	}
}

// Defines values for AddHarnessExecutable201JSONResponseBodyStatus.
const (
	Registered AddHarnessExecutable201JSONResponseBodyStatus = "registered"
)

// Valid indicates whether the value is a known member of the AddHarnessExecutable201JSONResponseBodyStatus enum.
func (e AddHarnessExecutable201JSONResponseBodyStatus) Valid() bool {
	switch e {
	case Registered:
		return true
	default:
		return false
	}
}

// Defines values for DisableHarnessExecutable200JSONResponseBodyEnabled.
const (
	False DisableHarnessExecutable200JSONResponseBodyEnabled = false
)

// Valid indicates whether the value is a known member of the DisableHarnessExecutable200JSONResponseBodyEnabled enum.
func (e DisableHarnessExecutable200JSONResponseBodyEnabled) Valid() bool {
	switch e {
	case False:
		return true
	default:
		return false
	}
}

// Defines values for EnableHarnessExecutable200JSONResponseBodyEnabled.
const (
	True EnableHarnessExecutable200JSONResponseBodyEnabled = true
)

// Valid indicates whether the value is a known member of the EnableHarnessExecutable200JSONResponseBodyEnabled enum.
func (e EnableHarnessExecutable200JSONResponseBodyEnabled) Valid() bool {
	switch e {
	case True:
		return true
	default:
		return false
	}
}

// CommandResponse defines model for CommandResponse.
type CommandResponse struct {
	EventId     string `json:"event_id"`
	Replayed    bool   `json:"replayed"`
	RoomId      string `json:"room_id"`
	RoomVersion int64  `json:"room_version"`
}

// Error defines model for Error.
type Error struct {
	Error struct {
		// Code 稳定错误码（transport 与 domain 映射）
		Code    string `json:"code"`
		Message string `json:"message"`
	} `json:"error"`
}

// Executable defines model for Executable.
type Executable struct {
	Adapter string `json:"adapter"`

	// Digest sha256（版本固定门禁用）
	Digest       *string    `json:"digest,omitempty"`
	DiscoveredAt *time.Time `json:"discovered_at,omitempty"`

	// Distro WSL 运行面的发行版（native 为空）
	Distro  *string `json:"distro,omitempty"`
	Enabled bool    `json:"enabled"`

	// Id 注册表唯一 ID（adapter@runtime:path）
	Id string `json:"id"`

	// LoginState logged_in | logged_out | unknown
	LoginState *string `json:"login_state,omitempty"`
	Path       string  `json:"path"`

	// Runtime native | wsl
	Runtime string `json:"runtime"`

	// Source auto_scan | manual
	Source  *string `json:"source,omitempty"`
	Version *string `json:"version,omitempty"`
}

// ManualExecutableRequest defines model for ManualExecutableRequest.
type ManualExecutableRequest struct {
	Adapter string  `json:"adapter"`
	Distro  *string `json:"distro,omitempty"`
	Path    string  `json:"path"`
	Runtime string  `json:"runtime"`
	Version *string `json:"version,omitempty"`
}

// RoomCommand 命令信封（RFC-0001）。payload 按 command_kind 严格校验（未知字段拒绝）：
// create_room → CreateRoomPayload；post_message → PostMessagePayload；
// pause_room → PauseRoomPayload；resume_room → 空 payload。
type RoomCommand struct {
	CommandKind         RoomCommandCommandKind `json:"command_kind"`
	ExpectedRoomVersion int                    `json:"expected_room_version"`

	// IdempotencyKey UUIDv7（服务端按 tenant+key+kind 去重；同键异指纹 409）
	IdempotencyKey string                 `json:"idempotency_key"`
	IssuedAt       time.Time              `json:"issued_at"`
	Payload        map[string]interface{} `json:"payload"`
}

// RoomCommandCommandKind defines model for RoomCommand.CommandKind.
type RoomCommandCommandKind string

// Snapshot defines model for Snapshot.
type Snapshot struct {
	AlgorithmVersion  int64          `json:"algorithm_version"`
	ProjectionVersion int64          `json:"projection_version"`
	RoomId            string         `json:"room_id"`
	RoomVersion       int64          `json:"room_version"`
	Timeline          []TimelineItem `json:"timeline"`
	Watermark         string         `json:"watermark"`
}

// TimelineItem defines model for TimelineItem.
type TimelineItem struct {
	ActorId    string    `json:"actor_id"`
	ActorKind  string    `json:"actor_kind"`
	Body       *string   `json:"body,omitempty"`
	EventId    string    `json:"event_id"`
	OccurredAt time.Time `json:"occurred_at"`
	Position   string    `json:"position"`
	ThreadId   *string   `json:"thread_id,omitempty"`
	Type       string    `json:"type"`
}

// ExecutableID defines model for ExecutableID.
type ExecutableID = string

// RoomID defines model for RoomID.
type RoomID = string

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Conflict defines model for Conflict.
type Conflict = Error

// HarnessUnavailable defines model for HarnessUnavailable.
type HarnessUnavailable = Error

// Internal defines model for Internal.
type Internal = Error

// NotFound defines model for NotFound.
type NotFound = Error

// OriginRejected defines model for OriginRejected.
type OriginRejected = Error

// TooLarge defines model for TooLarge.
type TooLarge = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// UnsupportedMediaType defines model for UnsupportedMediaType.
type UnsupportedMediaType = Error

// GetHealthz200JSONResponseBodyStatus defines parameters for GetHealthz.
type GetHealthz200JSONResponseBodyStatus string

// AddHarnessExecutable201JSONResponseBodyStatus defines parameters for AddHarnessExecutable.
type AddHarnessExecutable201JSONResponseBodyStatus string

// DisableHarnessExecutable200JSONResponseBodyEnabled defines parameters for DisableHarnessExecutable.
type DisableHarnessExecutable200JSONResponseBodyEnabled bool

// EnableHarnessExecutable200JSONResponseBodyEnabled defines parameters for EnableHarnessExecutable.
type EnableHarnessExecutable200JSONResponseBodyEnabled bool

// SubscribeRoomEventsParams defines parameters for SubscribeRoomEvents.
type SubscribeRoomEventsParams struct {
	// Cursor opaque 游标（空 = 从头）；与 Last-Event-ID 同义，query 优先
	Cursor *string `form:"cursor,omitempty" json:"cursor,omitempty"`
}

// AddHarnessExecutableJSONRequestBody defines body for AddHarnessExecutable for application/json ContentType.
type AddHarnessExecutableJSONRequestBody = ManualExecutableRequest

// CreateRoomJSONRequestBody defines body for CreateRoom for application/json ContentType.
type CreateRoomJSONRequestBody = RoomCommand

// SubmitRoomCommandJSONRequestBody defines body for SubmitRoomCommand for application/json ContentType.
type SubmitRoomCommandJSONRequestBody = RoomCommand

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetHealthz 存活探针
	// (GET /healthz)
	GetHealthz(w http.ResponseWriter, r *http.Request)
	// ListHarnessExecutables 宿主扫描到的 agent CLI 注册表
	// (GET /v1/harness/executables)
	ListHarnessExecutables(w http.ResponseWriter, r *http.Request)
	// AddHarnessExecutable 手动登记 agent CLI（探测失败拒收）
	// (POST /v1/harness/executables)
	AddHarnessExecutable(w http.ResponseWriter, r *http.Request)
	// DisableHarnessExecutable 禁用可执行项（座位随 resync 撤出）
	// (POST /v1/harness/executables/{id}/disable)
	DisableHarnessExecutable(w http.ResponseWriter, r *http.Request, id ExecutableID)
	// EnableHarnessExecutable 启用可执行项（登录门控；启用后 ≤10s 入座，无需重启）
	// (POST /v1/harness/executables/{id}/enable)
	EnableHarnessExecutable(w http.ResponseWriter, r *http.Request, id ExecutableID)
	// GetOwnerBootstrap 第一方客户端引导：返回 owner token（跨源门保护）
	// (GET /v1/owner/bootstrap)
	GetOwnerBootstrap(w http.ResponseWriter, r *http.Request)
	// CreateRoom 创建房间（create_room 命令）
	// (POST /v1/rooms)
	CreateRoom(w http.ResponseWriter, r *http.Request)
	// SubmitRoomCommand 提交房间命令（post_message / pause_room / resume_room）
	// (POST /v1/rooms/{room_id}/commands)
	SubmitRoomCommand(w http.ResponseWriter, r *http.Request, roomId RoomID)
	// SubscribeRoomEvents SSE 订阅房间事件流（opaque cursor 续传）
	// (GET /v1/rooms/{room_id}/events)
	SubscribeRoomEvents(w http.ResponseWriter, r *http.Request, roomId RoomID, params SubscribeRoomEventsParams)
	// GetRoomSnapshot 快照四元组（版本 + 水位 + 投影/算法版本 + Timeline 投影）
	// (GET /v1/rooms/{room_id}/snapshot)
	GetRoomSnapshot(w http.ResponseWriter, r *http.Request, roomId RoomID)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetHealthz operation middleware
func (siw *ServerInterfaceWrapper) GetHealthz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealthz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListHarnessExecutables operation middleware
func (siw *ServerInterfaceWrapper) ListHarnessExecutables(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListHarnessExecutables(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AddHarnessExecutable operation middleware
func (siw *ServerInterfaceWrapper) AddHarnessExecutable(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AddHarnessExecutable(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DisableHarnessExecutable operation middleware
func (siw *ServerInterfaceWrapper) DisableHarnessExecutable(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id ExecutableID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DisableHarnessExecutable(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// EnableHarnessExecutable operation middleware
func (siw *ServerInterfaceWrapper) EnableHarnessExecutable(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id ExecutableID

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.EnableHarnessExecutable(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetOwnerBootstrap operation middleware
func (siw *ServerInterfaceWrapper) GetOwnerBootstrap(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetOwnerBootstrap(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateRoom operation middleware
func (siw *ServerInterfaceWrapper) CreateRoom(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateRoom(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SubmitRoomCommand operation middleware
func (siw *ServerInterfaceWrapper) SubmitRoomCommand(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "room_id" -------------
	var roomId RoomID

	err = runtime.BindStyledParameterWithOptions("simple", "room_id", r.PathValue("room_id"), &roomId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "room_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SubmitRoomCommand(w, r, roomId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SubscribeRoomEvents operation middleware
func (siw *ServerInterfaceWrapper) SubscribeRoomEvents(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "room_id" -------------
	var roomId RoomID

	err = runtime.BindStyledParameterWithOptions("simple", "room_id", r.PathValue("room_id"), &roomId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "room_id", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params SubscribeRoomEventsParams

	// ------------- Optional query parameter "cursor" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "cursor", r.URL.Query(), &params.Cursor, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "cursor"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "cursor", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SubscribeRoomEvents(w, r, roomId, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetRoomSnapshot operation middleware
func (siw *ServerInterfaceWrapper) GetRoomSnapshot(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "room_id" -------------
	var roomId RoomID

	err = runtime.BindStyledParameterWithOptions("simple", "room_id", r.PathValue("room_id"), &roomId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "room_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRoomSnapshot(w, r, roomId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/healthz", wrapper.GetHealthz)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/v1/rooms", wrapper.CreateRoom)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/v1/rooms/{room_id}/commands", wrapper.SubmitRoomCommand)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/v1/rooms/{room_id}/events", wrapper.SubscribeRoomEvents)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/v1/rooms/{room_id}/snapshot", wrapper.GetRoomSnapshot)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/v1/owner/bootstrap", wrapper.GetOwnerBootstrap)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/v1/harness/executables", wrapper.ListHarnessExecutables)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/v1/harness/executables", wrapper.AddHarnessExecutable)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/v1/harness/executables/{id}/enable", wrapper.EnableHarnessExecutable)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/v1/harness/executables/{id}/disable", wrapper.DisableHarnessExecutable)

	return m
}